% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{dfw.beer.survey}
\alias{dfw.beer.survey}
\title{Dallas-Forth Worth beer and brewery survey (autumn/winter 2014)}
\format{A list with two items:
\describe{
\item{original}{A list with many items exported from Google Sheets. Contains information on respondents and beers for the survey.}
\item{mca.ex}{A list with three items that are the prepare data (for use with multiple correspondence analysis or similar technique)}
}}
\usage{
dfw.beer.survey
}
\description{
These data were created because Derek was bored and loves beer. See <http://www.dallasobserver.com/restaurants/dfws-favorite-local-beer-as-proven-by-an-actual-scientist-7024953> and <http://www.derekbeaton.com/dallass-favorite-beers/>
}
\keyword{datasets}
