% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExPositionData.R
\docType{package}
\name{ExPositionData-package}
\alias{ExPositionData}
\alias{ExPositionData-package}
\title{ExPositionData}
\description{
The purpose of this package is to house all of the data sets that have been curated for use with ExPosition. All packages in the ExPosition family (besides this one) will have only limited and necessary data sets. This has many data sets for use and exploration.
}
\author{
\strong{Maintainer}: Derek Beaton \email{exposition.software@gmail.com}

}
