% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{mcu.movie.actors}
\alias{mcu.movie.actors}
\title{Marvel Cinematic Universe (actors x movies; 2017).}
\format{A list with two items:
\describe{
\item{marvel_movies_actors.by.movies}{a 938 actors x 12806 movies matrix; values are 1 (actor was in a movie) or 0 (actor was not in a movie).}
\item{movie.info}{a 12806 movies x 2 attributes matrix; a simple list to map movie name (year) to the IMDB short codes.}
}}
\usage{
mcu.movie.actors
}
\description{
These data were created because Derek is a nerd. See <https://github.com/derekbeaton/Marvel-Cinematic-Universe_Network>
}
\keyword{datasets}
